packages/train-gui/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "train-gui"
version = "0.1.0"
description = "Interactive GUI for curating Gaussian component training sets from GRS spectra"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "matplotlib>=3.9",
    "numpy>=1.26",
    "astropy>=6.0",
    "scipy>=1.12",
    "benchmarks",
]

[project.scripts]
train-gui = "train_gui.cli:main"

[dependency-groups]
dev = ["pytest-cov>=7.0.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/train_gui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov --cov-report=term-missing"

[tool.coverage.run]
source = ["train_gui"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:", "\\.\\.\\."]
